refactor: make song index a usize to avoid casting on render
only casts in `rejitter_song`, which runs less often, and only one cast
in that file.
as for performance, this is probably more likely to slow it down*, but it
is less boilerplate in rendering.

*rust seems to do well with `u32`/`i32`, and the previous value was
`u16`, so it might improve performance**
**i dont know. i never tested

+++ src/tui.rs
@@ -26,7 +26,7 @@ pub struct Tui<'a> {
     width: u16,
     height: u16,
     scrolling_offset: usize,
-    pub cursor_index_queue: u16,
+    pub cursor_index_queue: usize,
 }
 
 impl Tui<'_> {
@@ -133,7 +133,7 @@ impl Tui<'_> {
         if self.cursor_index_queue as usize >= songs.len() {
             // wrap back to the size of songs; the user is trying to access songs.len() + 1
             // will panic otherwise, but callers dont need to care
-            self.cursor_index_queue = songs.len() as u16 - 1;
+            self.cursor_index_queue = songs.len() - 1;
             SONG_INDEX.store(self.cursor_index_queue, Relaxed);
         } else if self.scrolling_offset >= songs.len() {
             self.scrolling_offset = songs.len() - 1;
@@ -160,7 +160,7 @@ impl Tui<'_> {
         // HACK: for some reason, this code thinks cursor_index_queue^self.scrolling_offset is the
         // currently selected song. subtract it now.
         // i will give you a hug if you find out why that is, and a workaround that isn't this ugly.
-        self.cursor_index_queue -= self.scrolling_offset as u16;
+        self.cursor_index_queue -= self.scrolling_offset;
         for i in 0..(self.height as usize - 12) + self.scrolling_offset {
             if i > songs.len() {
                 break;
@@ -203,7 +203,7 @@ impl Tui<'_> {
     fn __draw_safe(&mut self) -> Result<(), std::io::Error> {
         let songs = &crate::PLAYLIST.read();
         if self.cursor_index_queue as usize >= songs.len() {
-            self.cursor_index_queue = songs.len() as u16 - 1;
+            self.cursor_index_queue = songs.len() - 1;
             SONG_INDEX.store(self.cursor_index_queue, Relaxed);
         }
         let song = songs[self.cursor_index_queue as usize].split("/").last().unwrap_or("");
